Let PestGuidance override ZestGuidance
